+++ CMakeLists.txt
@@ -8,7 +8,7 @@ else()
 endif()
 
 project(cpp-ap
-    VERSION 1.0
+    VERSION 2.0.0
     DESCRIPTION "Command-line argument parser for C++20"
     HOMEPAGE_URL "https://github.com/SpectraL519/cpp-ap"
     LANGUAGES CXX
